+++ src/Extension.fs
@@ -27,6 +27,7 @@ module private Headers =
 
 [<RequireQualifiedAccess>]
 module Http =
+    open System
     open Microsoft.AspNetCore.Http
 
     let private headersFromContext (httpContext: HttpContext): HeaderSeq =
@@ -36,14 +37,25 @@ module Http =
     let extractFromHeaders headers =
         let httpHeadersCarrier = TextMapExtractAdapter(headers |> headersToDictionary) :> ITextMap
 
-        match Tracer.tracer().Extract(BuiltinFormats.HttpHeaders, httpHeadersCarrier) with
-        | null -> Inactive
-        | context -> Context context
+        let trace =
+            match Tracer.tracer().Extract(BuiltinFormats.HttpHeaders, httpHeadersCarrier) with
+            | null -> Inactive
+            | context -> Context context
+
+        trace
+        |> Trace.addTags [
+            "component:", (sprintf "ftracing (%s)" AssemblyVersionInformation.AssemblyVersion)
+        ]
 
     let extractFromContext (httpContext: HttpContext) =
         httpContext
         |> headersFromContext
         |> extractFromHeaders
+        |> Trace.addTags [
+            "http.method", httpContext.Request.Method
+            "http.status_code", string httpContext.Response.StatusCode
+            "http.url", sprintf "%s://%s%s%s" httpContext.Request.Scheme httpContext.Request.Host.Value httpContext.Request.Path.Value httpContext.Request.QueryString.Value
+        ]
 
     let inject trace headers =
         match trace |> Trace.context with