+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Extract trace from http context with specific tags for HTTP
 
 ## 1.2.0 - 2021-07-26
 - Add Extension module with `Http` `inject/extract` functions